SyntheticTrajectoryGenerator/requirements.txt
adtk==0.6.2
aioredis==1.3.1
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.5.6
async-generator==1.10
async-timeout==3.0.1
asyncio-nats-client==0.11.4
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1599308529326/work
autobahn==20.7.1
Automat==20.2.0
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
backcall==0.2.0
bayesian-hmm==0.0.4
biopython==1.78
bleach==3.2.1
branca @ file:///home/conda/feedstock_root/build_artifacts/branca_1588860304135/work
brotlipy==0.7.0
cachetools==4.2.0
certifi==2021.5.30
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1625835293160/work
channels==2.4.0
channels-redis==3.1.0
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1600358500179/work
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
combo==0.1.1
ConfigArgParse==1.2.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1600960155296/work
cycler==0.10.0
Cython==0.29.21
danlp==0.0.9
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
descartes==1.1.0
Django==3.1.2
django-allow-healthchecks==0.1.4
django-cors-headers==3.5.0
django-filter==2.4.0
django-healthz==0.0.5
django-phonenumber-field==5.0.0
django-storages==1.10.1
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
docker==5.0.0
docopt==0.6.2
docutils==0.16
drf-nested-routers==0.91
drf-yasg==1.17.1
entrypoints==0.3
environs==9.3.1
extratools==0.8.2.1
Faker==4.4.0
filterpy==1.4.5
Fiona==1.8.17
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-CAS==1.0.2
Flask-Login==0.5.0
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
folium @ file:///home/conda/feedstock_root/build_artifacts/folium_1588861378954/work
future==0.18.2
GDAL==3.1.4
geohash-hilbert==1.4.0
geohash2==1.1
geojson==2.5.0
geopandas @ file:///home/conda/feedstock_root/build_artifacts/geopandas_1594925563431/work
gevent==20.9.0
geventhttpclient==1.4.4
google-auth==1.23.0
gpxpy==1.4.2
greenlet==0.4.17
gunicorn==20.0.4
h3 @ file:///home/conda/feedstock_root/build_artifacts/h3-py_1623580567120/work
hiredis==1.1.0
hmmlearn==0.2.5
huey==2.3.0
hyperlink==20.0.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
ikalman==0.2.0
incremental==17.5.0
inflection==0.5.1
infomap==1.3.0
infostop==0.1.8
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.8.0
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1593624380152/work
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.2.0
kubernetes @ git+https://github.com/kubernetes-client/python.git@818e5ba0ba8cfc6d866195de706f58b7431110c1
lazy-object-proxy==1.6.0
livereload==2.6.3
llvmlite==0.34.0
locust==1.4.1
lunr==0.5.8
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.10.0
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-base_1600159108024/work
mccabe==0.6.1
mistune==0.8.4
mkdocs==1.1.2
mmh3==2.5.1
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1612895720168/work
msgpack==1.0.0
munch==2.5.0
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1598210780226/work
nltk==3.5
notebook==6.1.4
numba==0.51.2
numpy==1.20.2
oauthlib==3.1.0
olefile==0.46
osmnx @ file:///home/conda/feedstock_root/build_artifacts/osmnx_1599521990085/work
packaging==20.4
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1599600959724/work
pandocfilters==1.4.2
parso==0.7.1
patsy==0.5.1
pexpect==4.8.0
phonenumbers==8.12.10
pickleshare==0.7.5
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1594213113401/work
plotly==4.14.1
pluggy==0.13.1
pomegranate==0.14.4
powerlaw @ file:///home/conda/feedstock_root/build_artifacts/powerlaw_1589637974211/work
prefixspan==0.5.2
prometheus-client==0.8.0
prompt-toolkit==3.0.7
psutil==5.7.3
psycopg2==2.8.5
ptyprocess==0.6.0
py==1.10.0
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pyconll==2.3.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
Pygments==2.7.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pykalman==0.9.5
pylint==2.8.3
pyod==0.8.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj @ file:///home/conda/feedstock_root/build_artifacts/pyproj_1623801867087/work
pyrsistent==0.17.3
pysad==0.1.1
PySocks==1.7.1
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==3.13.2
python-igraph @ file:///home/conda/feedstock_root/build_artifacts/python-igraph_1625644196590/work
python-louvain==0.15
python-socketio==4.6.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
rank-bm25==0.2.1
redis==3.5.3
redisxchange @ git+https://github.com/nicklasxyz/redisxchange@19696ca38688744921bd2465a407f1857b32e050
regex==2020.9.27
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1592425495151/work
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
Rtree==0.9.4
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1596546073704/work
scikit-mobility @ file:///home/conda/feedstock_root/build_artifacts/scikit-mobility_1617009946777/work
scikit-network==0.20.0
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1595583577416/work
seaborn==0.11.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
Shapely @ file:///home/conda/feedstock_root/build_artifacts/shapely_1611317103512/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
sortedcontainers==2.3.0
SQLAlchemy==1.3.20
sqlparse==0.3.1
statsmodels @ file:///home/conda/feedstock_root/build_artifacts/statsmodels_1612273599609/work
stumpy==1.5.0
suod==0.0.4
sympy==1.8
tabulate==0.8.9
terminado==0.9.1
terminaltables==3.1.0
testpath==0.4.4
text-unidecode==1.3
texttable @ file:///home/conda/feedstock_root/build_artifacts/texttable_1626204417032/work
threadpoolctl @ file:///tmp/tmp79xdzxkt/threadpoolctl-2.1.0-py3-none-any.whl
toml==0.10.2
toolz==0.11.1
tornado==6.0.4
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1625664838498/work
traitlets==5.0.4
tslearn==0.4.1
tweepy==3.9.0
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1595434816409/work
utm==0.7.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.3
xmltodict==0.12.0
zope.event==4.5.0
zope.interface==5.1.2
